refactor: move the Copilot SEA extraction into its own module
The 18 functions behind `ensure_copilot_extracted` already sat under one
cfg gate with their own fixture harness at the bottom of main.rs, so the
module existed in fact but not in file layout. Moving it to
src/copilot_extract.rs takes main.rs from 7266 to 6289 lines and leaves
the orchestration in run() readable without scrolling past a thousand
lines of cache-directory polling.

Pure relocation: the only edits are the `mod` declaration, the qualified
call site, and `pub` on the one entry point. The three extraction tests
that lived in main.rs's generic `tests` module move next to the code they
exercise. Test count is unchanged at 39 for the binary target.
